Move UDS.py back out of __init__
